+++ changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to `PHP CSV Iterator` will be documented in this file.
 
+## Version 1.0.2
+
+### Added/changed
+- Add function to get result as object
+- Add returns in docblocks
+
 ## Version 1.0.1
 
 ### Added/changed